extract routing primatives
