+++ src/lib/routes/slug.ts
@@ -1,0 +1,3 @@
+const slug = (s: string): string => s.replace(/\s/g, '-').toLowerCase();
+
+export default slug;
